% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{cat_function}
\alias{cat_function}
\title{Cat love function}
\usage{
cat_function(love = TRUE)
}
\arguments{
\item{love}{Enter TRUE if you love cats!}
}
\value{
A message
}
\description{
Cat love function
}
\examples{
cat_function(love = TRUE)
}
