% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_user}
\alias{hello_user}
\title{Greet a user}
\usage{
hello_user(name)
}
\arguments{
\item{name}{A String of any length, specifying the user that should be greeted.}
}
\description{
\code{hello_user} returns a personalized greeting, using the provided
\code{name} argument.
}
\details{
This function serves as an example for showing the different options for
documenting functions. Here you will find most of the \strong{@tags}
explained in the relevant chapter of the R package book found here:
\url{http://r-pkgs.had.co.nz}.
}
\section{Customized link text}{
 As a case in point the reference in the
  previous sentence could also have been written with customized link text,
  i.e. \href{http://r-pkgs.had.co.nz}{R package book}. And now we have also
  demonstrated the use of a \code{@section} tag.
}
\examples{
hello_user("A")
hello_user("B")

\dontrun{
hello_user(A)
}

}
\seealso{
\code{\link{hello}} which is in the same \code{@family} of
  functions, and of course check \url{http://r-pkgs.had.co.nz}.

Other greetings: \code{\link{hello}}
}

