% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IceCreamPriceFunction.R
\name{icecreampricefun}
\alias{icecreampricefun}
\title{Ice cream price function}
\usage{
icecreampricefun(scoops, cone, topping)
}
\arguments{
\item{scoops}{A number indicating price of scoops}

\item{cone}{A number indicating price of cone}

\item{topping}{A number indicating price of topping}
}
\value{
A number indicating the total price of your icecream
}
\description{
Ice cream price function
}
\examples{
icecreamprice(1,1,1)
}
