% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{numgen}
\alias{numgen}
\title{This function generates random numbers from a normal distribution.}
\usage{
numgen(n)
}
\arguments{
\item{n}{The number of random numbers to generate.}
}
\value{
A vector of random numbers.
}
\description{
This function generates random numbers from a normal distribution.
}
\examples{
numgen(10)
}
