% Generated by roxygen2 (4.0.0): do not edit by hand
\name{test}
\alias{test}
\title{test function}
\usage{
test(param1, param2)
}
\arguments{
  \item{param1}{first parameter of addition operation
  param2 first parameter of addition operation}
}
\description{
This function allows you to sum 2 numbers without the addition sign
}
\examples{
test(5,6)
test(-5,8)
}
\keyword{addition}
\keyword{summing}

