% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/this.R
\name{this}
\alias{this}
\title{A Test function}
\usage{
this(username = "David")
}
\arguments{
\item{username}{String. Name of user. Defaults to "David"}
}
\description{
This function is for R package development test.
Repeats hello 3 times.
}
\examples{
this
this(username="Erin")
}
\keyword{hello-world}
